+++ src/admin_ui/src/MappingField.tsx
@@ -3,7 +3,22 @@ import {Children, cloneElement, FunctionComponent, isValidElement, ReactElement}
 import {Record} from 'ra-core';
 import FormLabel from '@material-ui/core/FormLabel';
 import FormControl from '@material-ui/core/FormControl';
+import {Labeled} from 'react-admin';
+import {makeStyles} from '@material-ui/core/styles';
+import {Theme} from '@material-ui/core/styles/createMuiTheme';
 
+
+const useStyles = makeStyles((theme: Theme) => ({
+  fieldset: {
+      padding: '5px 12px',
+      border: `1px solid ${theme.palette.grey[500]}`,
+  },
+  legend: {
+      paddingLeft: '5px',
+      paddingRight: '5px',
+      fontSize: theme.typography.pxToRem(12),
+  },
+}));
 
 export interface MappingFieldProps {
     source: string;
@@ -25,9 +40,10 @@ export const MappingField: FunctionComponent<MappingFieldProps> =
          source,
          basePath,
      }) => {
+        const classes = useStyles();
         return (
-            <FormControl component="fieldset">
-                <FormLabel component="legend">{label}</FormLabel>
+            <FormControl component="fieldset" className={classes.fieldset}>
+                <FormLabel component="legend" className={classes.legend}>{label}</FormLabel>
                 {
                     Children.map(
                         children,
@@ -43,7 +59,7 @@ export const MappingField: FunctionComponent<MappingFieldProps> =
                                 field_index = undefined;
                             }
                             return (
-                                <>
+                                <Labeled label={field.props.label}>
                                     {cloneElement(field, {
                                         source: field_source,
                                         index: field_index,
@@ -51,7 +67,7 @@ export const MappingField: FunctionComponent<MappingFieldProps> =
                                         resource: resource,
                                         basePath: field.props.basePath || basePath,
                                     })}
-                                </>
+                                </Labeled>
                             );
                         }
                     )