+++ src/restfw_admin/CHANGES.rst
@@ -14,6 +14,14 @@
 CHANGELOG
 *********
 
+Next release
+============
+
+Changes
+-------
+
+- Added some CSS-styles for ``MappingField`` and ``MappingInput``.
+
 0.3 (2020-08-19)
 ================